requirements.txt
fastapi
uvicorn[standard]
pandas
scikit-learn
joblib
parselmouth
psycopg2-binary
SQLAlchemy
python-dotenv
numpy